+++ MountainCount/Assets.cs
@@ -1,6 +1,7 @@
 ﻿//using R2API.Utils;
 using RoR2;
 using System.Runtime.CompilerServices;
+using static MountainCount.Assets;
 
 namespace MountainCount
 {
@@ -17,6 +18,8 @@ namespace MountainCount
 
             [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.NoOptimization)]
             public abstract void GetCountExpanded(out object value1, out object value2, out object value3);
+
+            public abstract void ModifyShrineUseToken(ref Chat.SubjectFormatChatMessage subjectFormatChatMessage);
 
             [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.NoOptimization)]
             public virtual void AppendInfo()
@@ -53,6 +56,49 @@ namespace MountainCount
         {
             if (count == 0) return 0;
             return baseValue + stackValue * (count - 1);
+        }
+
+
+        public class ShrineMountain : ShrineReferenceBase
+        {
+            public override string SayCountToken => "MOUNTAINCOUNT_SAYAMOUNT_MOUNTAIN";
+
+            public override string AppendToken => "MOUNTAINCOUNT_SAYAMOUNT_MOUNTAIN_COMBINED";
+
+            public override int GetCount()
+            {
+                if (TeleporterInteraction.instance)
+                    return TeleporterInteraction.instance.shrineBonusStacks;
+                return 0;
+            }
+
+            public override void GetCountExpanded(out object expectedLimitedValue, out object expandedValue1, out object expandedValue2)
+            {
+                expectedLimitedValue = GetCount();
+                expandedValue1 = (int)expectedLimitedValue + 1;
+                if ((int)expectedLimitedValue == 0)
+                {
+                    expandedValue1 = 0;
+                }
+                expandedValue2 = -1;
+            }
+
+            public override void ModifyShrineUseToken(ref Chat.SubjectFormatChatMessage subjectFormatChatMessage)
+            {
+                subjectFormatChatMessage.baseToken = "MOUNTAINCOUNT_SHRINE_BOSS_USE_MESSAGE";
+                subjectFormatChatMessage.paramTokens = new string[] { GetCount().ToString() };
+            }
+
+            public override void SayCountExpanded()
+            {
+                GetCountExpanded(out object useCount, out object expandedCount, out object _);
+
+                RoR2.Chat.SendBroadcastChat(new RoR2.Chat.SimpleChatMessage()
+                {
+                    baseToken = SayCountExpandedToken,
+                    paramTokens = new string[] { useCount.ToString(), expandedCount.ToString() }
+                });
+            }
         }
     }
 }