+++ MountainCount/CHANGELOG.md
@@ -1,3 +1,7 @@
+* `0.2.01` - Dependency Fix
+	* Migrated MountainShrine reference to Assets
+	* Added a 'ModifyShrineUseToken' abstract method to ShrineReferenceBase in order to isolate `Chat_SendBroadcastChat_ChatMessageBase` method from trying to access fields from unloaded mod dlls
+	* Config default changes
 * `0.2.0` - Mod Compat
 	* Added support for Risk of Options.
 	* Added support for ExtraChallengeShrines